Bug 1858438 - Tabs Tray TestRail matching
