+++ fenix/app/src/androidTest/java/org/mozilla/fenix/ui/TabbedBrowsingTest.kt
@@ -7,20 +7,25 @@ package org.mozilla.fenix.ui
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.uiautomator.UiDevice
 import com.google.android.material.bottomsheet.BottomSheetBehavior
+import mozilla.components.browser.state.store.BrowserStore
+import mozilla.components.concept.engine.mediasession.MediaSession
 import okhttp3.mockwebserver.MockWebServer
 import org.junit.After
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.mozilla.fenix.customannotations.SmokeTest
+import org.mozilla.fenix.ext.components
 import org.mozilla.fenix.helpers.AndroidAssetDispatcher
 import org.mozilla.fenix.helpers.HomeActivityIntentTestRule
+import org.mozilla.fenix.helpers.MatcherHelper
 import org.mozilla.fenix.helpers.RetryTestRule
 import org.mozilla.fenix.helpers.TestAssetHelper
 import org.mozilla.fenix.helpers.TestHelper.closeApp
 import org.mozilla.fenix.helpers.TestHelper.restartApp
 import org.mozilla.fenix.helpers.TestHelper.verifyKeyboardVisibility
 import org.mozilla.fenix.ui.robots.browserScreen
+import org.mozilla.fenix.ui.robots.clickPageObject
 import org.mozilla.fenix.ui.robots.homeScreen
 import org.mozilla.fenix.ui.robots.navigationToolbar
 import org.mozilla.fenix.ui.robots.notificationShade
@@ -45,6 +50,7 @@ import org.mozilla.fenix.ui.robots.notificationShade
 class TabbedBrowsingTest {
     private lateinit var mDevice: UiDevice
     private lateinit var mockWebServer: MockWebServer
+    private lateinit var browserStore: BrowserStore
 
     @get:Rule
     val activityTestRule = HomeActivityIntentTestRule.withDefaultSettingsOverrides(skipOnboarding = true)
@@ -55,6 +61,10 @@ class TabbedBrowsingTest {
 
     @Before
     fun setUp() {
+        // Initializing this as part of class construction, below the rule would throw a NPE
+        // So we are initializing this here instead of in all related tests.
+        browserStore = activityTestRule.activity.components.core.store
+
         mDevice = UiDevice.getInstance(InstrumentationRegistry.getInstrumentation())
         mockWebServer = MockWebServer().apply {
             dispatcher = AndroidAssetDispatcher()
@@ -67,51 +77,7 @@ class TabbedBrowsingTest {
         mockWebServer.shutdown()
     }
 
-    @Test
-    fun openNewTabTest() {
-        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
-            mDevice.waitForIdle()
-            verifyTabCounter("1")
-        }.openTabDrawer {
-            verifyNormalModeSelected()
-            verifyExistingOpenTabs("Test_Page_1")
-            closeTab()
-        }.openTabDrawer {
-            verifyNoOpenTabsInNormalBrowsing()
-        }.openNewTab {
-        }.submitQuery(defaultWebPage.url.toString()) {
-            mDevice.waitForIdle()
-            verifyTabCounter("1")
-        }.openTabDrawer {
-            verifyNormalModeSelected()
-            verifyExistingOpenTabs("Test_Page_1")
-        }
-    }
-
-    // TestRail link: https://testrail.stage.mozaws.net/index.php?/cases/view/233629
-    @Test
-    fun openNewPrivateTabTest() {
-        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-
-        homeScreen {}.togglePrivateBrowsingMode()
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
-            mDevice.waitForIdle()
-            verifyTabCounter("1")
-        }.openTabDrawer {
-            verifyExistingTabList()
-            verifyPrivateModeSelected()
-        }.toggleToNormalTabs {
-            verifyNoOpenTabsInNormalBrowsing()
-        }.toggleToPrivateTabs {
-            verifyExistingTabList()
-        }
-    }
-
+    // TestRail: https://testrail.stage.mozaws.net/index.php?/cases/view/903599
     @Test
     fun closeAllTabsTest() {
         val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
@@ -144,45 +110,9 @@ class TabbedBrowsingTest {
         }
     }
 
+    // TestRail: https://testrail.stage.mozaws.net/index.php?/cases/view/903604
     @Test
-    fun closeTabTest() {
-        val genericURL = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(genericURL.url) {
-        }.openTabDrawer {
-            verifyExistingOpenTabs("Test_Page_1")
-            closeTab()
-        }
-        homeScreen {
-            verifyTabCounter("0")
-        }.openNavigationToolbar {
-        }.enterURLAndEnterToBrowser(genericURL.url) {
-        }.openTabDrawer {
-            verifyExistingOpenTabs("Test_Page_1")
-            swipeTabRight("Test_Page_1")
-        }
-        homeScreen {
-            verifyTabCounter("0")
-        }.openNavigationToolbar {
-        }.enterURLAndEnterToBrowser(genericURL.url) {
-        }.openTabDrawer {
-            verifyExistingOpenTabs("Test_Page_1")
-            swipeTabLeft("Test_Page_1")
-        }
-        homeScreen {
-            verifyTabCounter("0")
-        }
-    }
-
-    @Test
-    fun verifyUndoSnackBarTest() {
-        // disabling these features because they interfere with the snackbar visibility
-        activityTestRule.applySettingsExceptions {
-            it.isPocketEnabled = false
-            it.isRecentTabsFeatureEnabled = false
-        }
-
+    fun closingTabsMethodsTest() {
         val genericURL = TestAssetHelper.getGenericAsset(mockWebServer, 1)
 
         navigationToolbar {
@@ -193,24 +123,9 @@ class TabbedBrowsingTest {
             verifySnackBarText("Tab closed")
             snackBarButtonClick("UNDO")
         }
-
         browserScreen {
             verifyTabCounter("1")
         }.openTabDrawer {
-            verifyExistingOpenTabs("Test_Page_1")
-        }
-    }
-
-    @Test
-    fun closePrivateTabTest() {
-        val genericURL = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-
-        homeScreen { }.togglePrivateBrowsingMode()
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(genericURL.url) {
-        }.openTabDrawer {
-            verifyExistingOpenTabs("Test_Page_1")
-            verifyCloseTabsButton("Test_Page_1")
             closeTab()
         }
         homeScreen {
@@ -220,6 +135,7 @@ class TabbedBrowsingTest {
         }.openTabDrawer {
             verifyExistingOpenTabs("Test_Page_1")
             swipeTabRight("Test_Page_1")
+            verifySnackBarText("Tab closed")
         }
         homeScreen {
             verifyTabCounter("0")
@@ -228,14 +144,16 @@ class TabbedBrowsingTest {
         }.openTabDrawer {
             verifyExistingOpenTabs("Test_Page_1")
             swipeTabLeft("Test_Page_1")
+            verifySnackBarText("Tab closed")
         }
         homeScreen {
             verifyTabCounter("0")
         }
     }
 
+    // TestRail: https://testrail.stage.mozaws.net/index.php?/cases/view/903591
     @Test
-    fun verifyPrivateTabUndoSnackBarTest() {
+    fun closingPrivateTabsMethodsTest() {
         val genericURL = TestAssetHelper.getGenericAsset(mockWebServer, 1)
 
         homeScreen { }.togglePrivateBrowsingMode()
@@ -248,12 +166,51 @@ class TabbedBrowsingTest {
             verifySnackBarText("Private tab closed")
             snackBarButtonClick("UNDO")
         }
-
         browserScreen {
             verifyTabCounter("1")
         }.openTabDrawer {
+            closeTab()
+        }
+        homeScreen {
+            verifyTabCounter("0")
+        }.openNavigationToolbar {
+        }.enterURLAndEnterToBrowser(genericURL.url) {
+        }.openTabDrawer {
             verifyExistingOpenTabs("Test_Page_1")
-            verifyPrivateModeSelected()
+            swipeTabRight("Test_Page_1")
+            verifySnackBarText("Private tab closed")
+        }
+        homeScreen {
+            verifyTabCounter("0")
+        }.openNavigationToolbar {
+        }.enterURLAndEnterToBrowser(genericURL.url) {
+        }.openTabDrawer {
+            verifyExistingOpenTabs("Test_Page_1")
+            swipeTabLeft("Test_Page_1")
+            verifySnackBarText("Private tab closed")
+        }
+        homeScreen {
+            verifyTabCounter("0")
+        }
+    }
+
+    // TestRail: https://testrail.stage.mozaws.net/index.php?/cases/view/903606
+    @SmokeTest
+    @Test
+    fun tabMediaControlButtonTest() {
+        val audioTestPage = TestAssetHelper.getAudioPageAsset(mockWebServer)
+
+        navigationToolbar {
+        }.enterURLAndEnterToBrowser(audioTestPage.url) {
+            mDevice.waitForIdle()
+            clickPageObject(MatcherHelper.itemWithText("Play"))
+            assertPlaybackState(browserStore, MediaSession.PlaybackState.PLAYING)
+        }.openTabDrawer {
+            verifyTabMediaControlButtonState("Pause")
+            clickTabMediaControlButton("Pause")
+            verifyTabMediaControlButtonState("Play")
+        }.openTab(audioTestPage.title) {
+            assertPlaybackState(browserStore, MediaSession.PlaybackState.PAUSED)
         }
     }
 
@@ -275,6 +232,40 @@ class TabbedBrowsingTest {
             verifyPrivateTabsNotification()
         }.clickClosePrivateTabsNotification {
             verifyHomeScreen()
+        }
+    }
+
+    // TestRail: https://testrail.stage.mozaws.net/index.php?/cases/view/903598
+    @SmokeTest
+    @Test
+    fun shareTabsFromTabsTrayTest() {
+        val firstWebsite = TestAssetHelper.getGenericAsset(mockWebServer, 1)
+        val secondWebsite = TestAssetHelper.getGenericAsset(mockWebServer, 2)
+        val firstWebsiteTitle = firstWebsite.title
+        val secondWebsiteTitle = secondWebsite.title
+        val sharingApp = "Gmail"
+        val sharedUrlsString = "${firstWebsite.url}\n\n${secondWebsite.url}"
+
+        homeScreen {
+        }.openNavigationToolbar {
+        }.enterURLAndEnterToBrowser(firstWebsite.url) {
+            verifyPageContent(firstWebsite.content)
+        }.openTabDrawer {
+        }.openNewTab {
+        }.submitQuery(secondWebsite.url.toString()) {
+            verifyPageContent(secondWebsite.content)
+        }.openTabDrawer {
+            verifyExistingOpenTabs("Test_Page_1")
+            verifyExistingOpenTabs("Test_Page_2")
+        }.openTabsListThreeDotMenu {
+            verifyShareAllTabsButton()
+        }.clickShareAllTabsButton {
+            verifyShareTabsOverlay(firstWebsiteTitle, secondWebsiteTitle)
+            verifySharingWithSelectedApp(
+                sharingApp,
+                sharedUrlsString,
+                "$firstWebsiteTitle, $secondWebsiteTitle",
+            )
         }
     }
 
@@ -301,6 +292,7 @@ class TabbedBrowsingTest {
         }
     }
 
+    // TestRail: https://testrail.stage.mozaws.net/index.php?/cases/view/903600
     @Test
     fun verifyEmptyTabTray() {
         navigationToolbar {
@@ -315,8 +307,9 @@ class TabbedBrowsingTest {
         }
     }
 
+    // TestRail: https://testrail.stage.mozaws.net/index.php?/cases/view/903585
     @Test
-    fun emptyTabsTrayViewPrivateBrowsingTest() {
+    fun verifyEmptyPrivateTabsTrayTest() {
         navigationToolbar {
         }.openTabTray {
         }.toggleToPrivateTabs {
@@ -330,12 +323,15 @@ class TabbedBrowsingTest {
         }
     }
 
+    // TestRail: https://testrail.stage.mozaws.net/index.php?/cases/view/903601
     @Test
-    fun verifyOpenTabDetails() {
+    fun verifyTabsTrayWithOpenTabTest() {
         val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
 
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
+        homeScreen {
+        }.openTabDrawer {
+        }.openNewTab {
+        }.submitQuery(defaultWebPage.url.toString()) {
         }.openTabDrawer {
             verifyNormalBrowsingButtonIsSelected(true)
             verifyPrivateBrowsingButtonIsSelected(false)
@@ -353,29 +349,57 @@ class TabbedBrowsingTest {
         }
     }
 
+    // TestRail: https://testrail.stage.mozaws.net/index.php?/cases/view/903587
+    @SmokeTest
     @Test
-    fun verifyContextMenuShortcuts() {
+    fun verifyPrivateTabsTrayWithOpenTabTest() {
+        val website = TestAssetHelper.getGenericAsset(mockWebServer, 1)
+
+        homeScreen {
+        }.openTabDrawer {
+        }.toggleToPrivateTabs {
+        }.openNewTab {
+        }.submitQuery(website.url.toString()) {
+        }.openTabDrawer {
+            verifyNormalBrowsingButtonIsSelected(false)
+            verifyPrivateBrowsingButtonIsSelected(true)
+            verifySyncedTabsButtonIsSelected(false)
+            verifyTabTrayOverflowMenu(true)
+            verifyTabsTrayCounter()
+            verifyExistingTabList()
+            verifyExistingOpenTabs(website.title)
+            verifyCloseTabsButton(website.title)
+            verifyOpenedTabThumbnail()
+            verifyPrivateBrowsingNewTabButton()
+        }
+    }
+
+    // TestRail: https://testrail.stage.mozaws.net/index.php?/cases/view/927315
+    @Test
+    fun tabsCounterShortcutMenuTest() {
         val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
 
         navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
+        }.enterURLAndEnterToBrowser(defaultWebPage.url) {}
+        navigationToolbar {
         }.openTabButtonShortcutsMenu {
             verifyTabButtonShortcutMenuItems()
         }.closeTabFromShortcutsMenu {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
+        }.enterURLAndEnterToBrowser(defaultWebPage.url) {}
+        navigationToolbar {
         }.openTabButtonShortcutsMenu {
         }.openNewPrivateTabFromShortcutsMenu {
             verifyKeyboardVisibility()
             verifySearchBarPlaceholder("Search or enter address")
             // dismiss search dialog
         }.dismissSearchBar {
-            verifyCommonMythsLink()
-            verifyNavigationToolbar()
+            verifyPrivateBrowsingHomeScreenItems()
         }
         navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
+        }.enterURLAndEnterToBrowser(defaultWebPage.url) {}
+        navigationToolbar {
         }.openTabButtonShortcutsMenu {
-        }.openTabFromShortcutsMenu {
+        }.openNewTabFromShortcutsMenu {
             verifyKeyboardVisibility()
             verifySearchBarPlaceholder("Search or enter address")
             // dismiss search dialog
@@ -385,6 +409,45 @@ class TabbedBrowsingTest {
         }
     }
 
+    // TestRail: https://testrail.stage.mozaws.net/index.php?/cases/view/927314
+    @Test
+    fun privateTabsCounterShortcutMenuTest() {
+        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
+
+        homeScreen {}.togglePrivateBrowsingMode()
+        navigationToolbar {
+        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
+            waitForPageToLoad()
+        }
+        navigationToolbar {
+        }.openTabButtonShortcutsMenu {
+            verifyTabButtonShortcutMenuItems()
+        }.closeTabFromShortcutsMenu {
+        }.enterURLAndEnterToBrowser(defaultWebPage.url) {}
+        navigationToolbar {
+        }.openTabButtonShortcutsMenu {
+        }.openNewPrivateTabFromShortcutsMenu {
+            verifyKeyboardVisibility()
+            verifySearchBarPlaceholder("Search or enter address")
+            // dismiss search dialog
+        }.dismissSearchBar {
+            verifyCommonMythsLink()
+        }
+        navigationToolbar {
+        }.enterURLAndEnterToBrowser(defaultWebPage.url) {}
+        navigationToolbar {
+        }.openTabButtonShortcutsMenu {
+        }.openNewTabFromShortcutsMenu {
+            verifyKeyboardVisibility()
+            verifySearchBarPlaceholder("Search or enter address")
+            // dismiss search dialog
+        }.dismissSearchBar {
+            // Verify normal browsing homescreen
+            verifyExistingTopSitesList()
+        }
+    }
+
+    // TestRail: https://testrail.stage.mozaws.net/index.php?/cases/view/1046683
     @Test
     fun verifySyncedTabsWhenUserIsNotSignedInTest() {
         navigationToolbar {
@@ -411,7 +474,7 @@ class TabbedBrowsingTest {
         closeApp(activityTestRule)
         restartApp(activityTestRule)
         homeScreen {
-            verifyPrivateBrowsingHomeScreen()
+            verifyPrivateBrowsingHomeScreenItems()
         }.openTabDrawer {
         }.toggleToNormalTabs {
             verifyExistingOpenTabs(defaultWebPage.title)
@@ -439,7 +502,7 @@ class TabbedBrowsingTest {
         restartApp(activityTestRule)
 
         homeScreen {
-            verifyPrivateBrowsingHomeScreen()
+            verifyPrivateBrowsingHomeScreenItems()
         }.openTabDrawer {
             verifyNoOpenTabsInPrivateBrowsing()
         }
