+++ fenix/app/src/androidTest/java/org/mozilla/fenix/ui/robots/HomeScreenRobot.kt
@@ -87,11 +87,10 @@ class HomeScreenRobot {
 
     fun verifyHomeScreen() = assertItemWithResIdExists(homeScreen)
 
-    fun verifyPrivateBrowsingHomeScreen() {
+    fun verifyPrivateBrowsingHomeScreenItems() {
         verifyHomeScreenAppBarItems()
         assertItemContainingTextExists(itemContainingText(privateSessionMessage))
         verifyCommonMythsLink()
-        verifyNavigationToolbarItems()
     }
 
     fun verifyHomeScreenAppBarItems() =
